Rename yocs_cmd_vel_mux to cmd_vel_mux.
