+++ package.xml
@@ -1,5 +1,5 @@
 <package>
-  <name>yocs_cmd_vel_mux</name>
+  <name>cmd_vel_mux</name>
   <version>0.12.1</version>
   <description>
      A multiplexer for command velocity inputs. Arbitrates incoming cmd_vel messages from several topics,
@@ -13,7 +13,7 @@
   <url type="website">http://ros.org/wiki/yocs_cmd_vel_mux</url>
   <url type="repository">https://github.com/yujinrobot/yujin_ocs</url>
   <url type="bugtracker">https://github.com/yujinrobot/yujin_ocs/issues</url>
-  
+
   <buildtool_depend>catkin</buildtool_depend>
 
   <build_depend>roscpp</build_depend>
